monitoring-system/log-parser/requirements.txt
kafka-python
kubernetes